code/requirements.txt
transformers~=4.26
Pillow~=9.4.0
requests~=2.28.0
